# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 班级网站网站 API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /admin/daily-practice/test-access:
    get:
      summary: 获取管理员专用内部测试用户开关
      x-required-roles: [ADMIN]
      responses:
        "200":
          description: 总开关状态、设置修订号和已选测试用户ID；计划暂停仍有效
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeTestAccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/daily-practice/test-access/{userId}:
    patch:
      summary: 仅在总开关关闭时设置单个用户的内部测试权限
      x-required-roles: [ADMIN]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled, expectedRevision]
              properties:
                enabled: { type: boolean }
                expectedRevision: { type: integer, minimum: 0 }
      responses:
        "200":
          description: 更新后的测试用户配置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeTestAccess" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/daily-practice/settings:
    get:
      summary: 获取每日一练总开关及当前有效服务状态
      x-required-roles: [EDITOR, ADMIN]
      responses:
        "200":
          description: 每日一练服务配置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      summary: 按修订号开启或关闭每日一练
      x-required-roles: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DailyPracticeSettingsUpdateRequest" }
      responses:
        "200":
          description: 更新后的服务配置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeSettings" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/daily-practice/service-pauses:
    get:
      summary: 分页获取假期停服与取消记录
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 停服时段列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeServicePausePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: 创建立即或未来生效的停服时段
      x-required-roles: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DailyPracticeServicePauseCreateRequest" }
      responses:
        "201":
          description: 停服时段已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeServicePause" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/daily-practice/service-pauses/{id}/cancel:
    post:
      summary: 提前恢复或取消未来停服时段
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: 停服记录已标记取消
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeServicePause" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/daily-practice/curriculum:
    get:
      summary: "读取两门专业课的教学日历、已学主题和题目匹配覆盖率"
      x-required-roles: ["EDITOR", "ADMIN"]
      responses:
        "200":
          description: "操作结果"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeCurriculumResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/daily-practice/curriculum/initialize:
    post:
      summary: "幂等初始化课表中的共同理论课进度，不覆盖已有修订"
      x-required-roles: ["EDITOR", "ADMIN"]
      responses:
        "201":
          description: "操作结果"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeCurriculumResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/daily-practice/curriculum/{id}:
    patch:
      summary: "按修订号更新主题进度、暂停状态或考试日期，并使未开始计划失效"
      x-required-roles: ["EDITOR", "ADMIN"]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeCourseUpdateRequest"
      responses:
        "200":
          description: "操作结果"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeCourseView"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/daily-practice/curriculum/{id}/history:
    get:
      summary: "读取课程最近一百次不可变修订"
      x-required-roles: ["EDITOR", "ADMIN"]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: "操作结果"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeCourseHistoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/daily-practice/teaching-progress:
    get:
      summary: 只读分页获取旧知识节点教学进度版本
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: subjectId
          schema: { type: string }
      responses:
        "200":
          description: 教学进度版本列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeachingProgressPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/daily-practice/teaching-progress/{id}:
    get:
      summary: 获取教学进度版本及未解析节点报告
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 教学进度详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeachingProgressDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/daily-practice/fixed-question-candidates:
    get:
      summary: 按目标练习日筛选两门课程已学理论主题中的 MANUAL/CSV 固定题候选
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: query
          name: practiceDate
          description: 目标练习日；省略时使用上海时区下一练习日，按目标日已学主题和课程考试日期筛选
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: subjectId
          schema: { type: string }
        - in: query
          name: chapterIds
          schema: { type: string, description: 逗号分隔章节 ID }
        - in: query
          name: chapterMatch
          schema: { type: string, enum: [ANY, ALL], default: ANY }
        - in: query
          name: includeCrossChapter
          schema: { type: boolean, default: false }
        - in: query
          name: gradingType
          schema: { $ref: "#/components/schemas/QuestionType" }
        - in: query
          name: typeLabel
          schema: { type: string, maxLength: 100 }
        - in: query
          name: pastPaper
          schema: { type: string, enum: [ALL, EXCLUDE, ONLY] }
        - in: query
          name: paperId
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string, maxLength: 200 }
      responses:
        "200":
          description: 非 AI 固定题候选
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeFixedQuestionPage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/daily-practice/fixed-assignments/{practiceDate}:
    get:
      summary: 获取某练习日最新固定附加题修订
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: practiceDate
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: 最新修订；从未发布时 assignment 为 null 且 revision 为 0
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeFixedAssignmentEnvelope" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/daily-practice/fixed-assignments:
    post:
      summary: 发布未来练习日的 0-20 道全班固定附加题
      x-required-roles: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DailyPracticeFixedAssignmentPublishRequest" }
      responses:
        "201":
          description: 固定题修订已发布
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeFixedAssignment" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/daily-practice/cycles:
    get:
      summary: 分页获取每日练习周期摘要（默认近 14 个练习日）
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: 周期摘要列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeCyclePage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/daily-practice/cycles/{practiceDate}:
    get:
      summary: 获取不含用户身份的每日周期聚合运行状态
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: practiceDate
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: 匿名周期聚合
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeCycleAggregate" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/daily-practice/cycles/{practiceDate}/refreeze:
    post:
      summary: 请求按当前教学进度与题库状态重新冻结当前练习日周期并重建未开始计划（Worker 异步执行）
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: practiceDate
          required: true
          schema: { type: string, format: date }
      responses:
        "201":
          description: 重新冻结已排队；alreadyRequested 表示此前已有未执行的请求
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyPracticeCycleRefreezeResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bmc3_session
  parameters:
...
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
...
  responses:
...
    Unauthorized:
      description: 未登录、会话失效或凭据无效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: 当前账号无权限或操作被安全规则拒绝
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: 目标不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: 操作与当前资源状态冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    Subject:
      type: object
      additionalProperties: false
      required: [id, name, slug, sortOrder, active]
      properties:
        id: { type: string }
        name: { type: string, maxLength: 100 }
        slug: { type: string, maxLength: 100 }
        sortOrder: { type: integer }
        active: { type: boolean }
...
    ApiError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        code: { type: string }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        requestId: { type: string }
    Author:
      type: object
      required: [id, displayName]
      properties:
        id: { type: string }
        displayName: { type: string }
...
    QuestionType:
      type: string
      enum: [SINGLE, MULTIPLE, TRUE_FALSE, SHORT_ANSWER]
...
    DailyPracticeCycleStatus:
      type: string
      enum: [BUILDING, GENERATING, READY, PAUSED, DEGRADED, FAILED]
...
    DailyPracticeServiceState:
      type: object
      additionalProperties: false
      required: [enabled, paused, reason, resumesAt, settingsRevision]
      properties:
        enabled: { type: boolean }
        paused: { type: boolean }
        reason: { oneOf: [{ type: string }, { type: "null" }] }
        resumesAt: { oneOf: [{ type: string, format: date-time }, { type: "null" }] }
        settingsRevision: { type: integer, minimum: 0 }
...
    DailyPracticeTestAccess:
      type: object
      additionalProperties: false
      required: [globalEnabled, revision, userIds]
      properties:
        globalEnabled: { type: boolean }
        revision: { type: integer, minimum: 0 }
        userIds:
          type: array
          uniqueItems: true
          items: { type: string }
    DailyPracticeSettings:
      type: object
      additionalProperties: false
      required: [enabled, revision, updatedAt, updatedBy, effective]
      properties:
        enabled: { type: boolean }
        revision: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
        updatedBy:
          oneOf:
            - { $ref: "#/components/schemas/Author" }
            - { type: "null" }
        effective: { $ref: "#/components/schemas/DailyPracticeServiceState" }
    DailyPracticeSettingsUpdateRequest:
      type: object
      additionalProperties: false
      required: [enabled, expectedRevision, reason]
      properties:
        enabled: { type: boolean }
        expectedRevision: { type: integer, minimum: 0 }
        reason: { type: string, minLength: 1, maxLength: 200 }
    DailyPracticeServicePause:
      type: object
      additionalProperties: false
      required: [id, startsAt, endsAt, reason, cancelledAt, createdAt, createdBy, cancelledBy]
      properties:
        id: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        reason: { type: string }
        cancelledAt: { oneOf: [{ type: string, format: date-time }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        createdBy:
          oneOf:
            - { $ref: "#/components/schemas/Author" }
            - { type: "null" }
        cancelledBy:
          oneOf:
            - { $ref: "#/components/schemas/Author" }
            - { type: "null" }
    DailyPracticeServicePauseCreateRequest:
      type: object
      additionalProperties: false
      required: [startsAt, endsAt, reason]
      properties:
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        reason: { type: string, minLength: 1, maxLength: 200 }
    DailyPracticeServicePausePage:
      type: object
      additionalProperties: false
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DailyPracticeServicePause" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
...
    TeachingProgressSummary:
      type: object
      required: [id, subject, version, effectivePracticeDate, changeType, note, correctionReason, scopeHash, nodeCount, publishedBy, publishedAt]
      properties:
        id: { type: string }
        subject: { $ref: "#/components/schemas/Subject" }
        version: { type: integer, minimum: 1 }
        effectivePracticeDate: { type: string, format: date }
        changeType: { $ref: "#/components/schemas/TeachingProgressChangeType" }
        note: { oneOf: [{ type: string }, { type: "null" }] }
        correctionReason: { oneOf: [{ type: string }, { type: "null" }] }
        scopeHash: { type: string, minLength: 64, maxLength: 64 }
        nodeCount: { type: integer, minimum: 0 }
        publishedBy:
          oneOf:
            - { $ref: "#/components/schemas/Author" }
            - { type: "null" }
        publishedAt: { type: string, format: date-time }
    TeachingProgressDetail:
      allOf:
        - { $ref: "#/components/schemas/TeachingProgressSummary" }
        - type: object
          required: [basedOnProgressId, nodes, unresolvedNodeCount, eligibleQuestionCount]
          properties:
            basedOnProgressId: { oneOf: [{ type: string }, { type: "null" }] }
            nodes:
              type: array
              maxItems: 5000
              items: { $ref: "#/components/schemas/TeachingProgressNodeSnapshot" }
            unresolvedNodeCount: { type: integer, minimum: 0 }
            eligibleQuestionCount: { type: integer, minimum: 0 }
    TeachingProgressPage:
      type: object
      additionalProperties: false
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TeachingProgressSummary" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
...
    DailyPracticeFixedQuestionPage:
      type: object
      additionalProperties: false
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DailyPracticeFixedQuestion" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
    DailyPracticeFixedAssignment:
      type: object
      additionalProperties: false
      required: [id, practiceDate, revision, basedOnAssignmentId, assignmentHash, note, publishedAt, publishedBy, questions]
      properties:
        id: { type: string }
        practiceDate: { type: string, format: date }
        revision: { type: integer, minimum: 1 }
        basedOnAssignmentId: { oneOf: [{ type: string }, { type: "null" }] }
        assignmentHash: { type: string, minLength: 64, maxLength: 64 }
        note: { oneOf: [{ type: string }, { type: "null" }] }
        publishedAt: { type: string, format: date-time }
        publishedBy:
          oneOf:
            - { $ref: "#/components/schemas/Author" }
            - { type: "null" }
        questions:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/DailyPracticeFixedQuestion" }
    DailyPracticeFixedAssignmentEnvelope:
      type: object
      additionalProperties: false
      required: [id, practiceDate, revision, basedOnAssignmentId, assignmentHash, note, publishedAt, publishedBy, questions, history, historyTotal]
      properties:
        id: { type: string }
        practiceDate: { type: string, format: date }
        revision: { type: integer, minimum: 1 }
        basedOnAssignmentId: { oneOf: [{ type: string }, { type: "null" }] }
        assignmentHash: { type: string, minLength: 64, maxLength: 64 }
        note: { oneOf: [{ type: string }, { type: "null" }] }
        publishedAt: { type: string, format: date-time }
        publishedBy:
          oneOf:
            - { $ref: "#/components/schemas/Author" }
            - { type: "null" }
        questions:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/DailyPracticeFixedQuestion" }
        history:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/DailyPracticeFixedAssignment" }
        historyTotal: { type: integer, minimum: 0 }
    DailyPracticeFixedAssignmentPublishRequest:
      type: object
      additionalProperties: false
      required: [practiceDate, expectedRevision, questionIds]
      properties:
        practiceDate: { type: string, format: date }
        expectedRevision: { type: integer, minimum: 0 }
        questionIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 191 }
        note: { type: string, minLength: 1, maxLength: 1000 }
...
    DailyPracticeCyclePage:
      type: object
      additionalProperties: false
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DailyPracticeCycleSummary" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
    DailyPracticeCycleRefreezeResponse:
      type: object
      additionalProperties: false
      required: [practiceDate, status, refreezeRequestedAt, rebuildableDayCount, alreadyRequested]
      properties:
        practiceDate: { type: string, format: date }
        status: { $ref: "#/components/schemas/DailyPracticeCycleStatus" }
        refreezeRequestedAt: { oneOf: [{ type: string, format: date-time }, { type: "null" }] }
        rebuildableDayCount: { type: integer, minimum: 0 }
        alreadyRequested: { type: boolean }
...
    PracticeCourseTopic:
      type: "object"
      additionalProperties: false
      required: ["id", "title", "sessionDates", "sourceRefs", "paused", "taughtOnOverride"]
      properties:
        id:
          type: "string"
          pattern: "^[a-z0-9-]{1,80}$"
        title:
          type: "string"
          minLength: 1
          maxLength: 200
        sessionDates:
          type: "array"
          items:
            type: "string"
            format: "date"
          minItems: 1
          maxItems: 100
        sourceRefs:
          type: "array"
          items:
            type: "string"
          maxItems: 100
        paused:
          type: "boolean"
        taughtOnOverride:
          type: ["string", "null"]
          format: "date"
    PracticeCourseView:
      type: "object"
      additionalProperties: false
      required: ["id", "subjectId", "subject", "termKey", "startDate", "examDate", "enabled", "revision", "topicHash", "topics", "status", "eligibleQuestionCount", "mappingCounts"]
      properties:
        id:
          type: "string"
        subjectId:
          type: "string"
        subject:
          type: "object"
          additionalProperties: false
          required: ["id", "name", "slug"]
          properties:
            id:
              type: "string"
            name:
              type: "string"
            slug:
              type: "string"
        termKey:
          type: "string"
        startDate:
          type: "string"
          format: "date"
        examDate:
          type: "string"
          format: "date"
        enabled:
          type: "boolean"
        revision:
          type: "integer"
          minimum: 1
          maximum: 2147483647
        topicHash:
          type: "string"
        topics:
          type: "array"
          items:
            type: "object"
            additionalProperties: false
            required: ["id", "title", "sessionDates", "sourceRefs", "paused", "taughtOnOverride", "taughtOn", "availableOn", "learned"]
            properties:
              id:
                type: "string"
                pattern: "^[a-z0-9-]{1,80}$"
              title:
                type: "string"
                minLength: 1
                maxLength: 200
              sessionDates:
                type: "array"
                items:
                  type: "string"
                  format: "date"
                minItems: 1
                maxItems: 100
              sourceRefs:
                type: "array"
                items:
                  type: "string"
                maxItems: 100
              paused:
                type: "boolean"
              taughtOnOverride:
                type: ["string", "null"]
                format: "date"
              taughtOn:
                type: "string"
                format: "date"
              availableOn:
                type: "string"
                format: "date"
              learned:
                type: "boolean"
          maxItems: 300
        status:
          type: "string"
          enum: ["UPCOMING", "ACTIVE", "COMPLETED", "PAUSED"]
        eligibleQuestionCount:
          type: "integer"
          minimum: 0
        mappingCounts:
          type: "object"
          additionalProperties: false
          required: ["PENDING", "PROCESSING", "READY", "NEEDS_REVIEW", "STALE"]
          properties:
            PENDING:
              type: "integer"
              minimum: 0
            PROCESSING:
              type: "integer"
              minimum: 0
            READY:
              type: "integer"
              minimum: 0
            NEEDS_REVIEW:
              type: "integer"
              minimum: 0
            STALE:
              type: "integer"
              minimum: 0
    PracticeCurriculumResponse:
      type: "object"
      additionalProperties: false
      required: ["practiceDate", "initialized", "courses"]
      properties:
        practiceDate:
          type: "string"
          format: "date"
        initialized:
          type: "boolean"
        courses:
          type: "array"
          items:
            $ref: "#/components/schemas/PracticeCourseView"
    PracticeCourseUpdateRequest:
      type: "object"
      additionalProperties: false
      required: ["expectedRevision", "reason"]
      properties:
        expectedRevision:
          type: "integer"
          minimum: 1
          maximum: 2147483647
        examDate:
          type: "string"
          format: "date"
        enabled:
          type: "boolean"
        topics:
          type: "array"
          items:
            $ref: "#/components/schemas/PracticeCourseTopic"
          minItems: 1
          maxItems: 300
        reason:
          type: "string"
          minLength: 1
          maxLength: 1000
    PracticeCourseHistoryResponse:
      type: "object"
      additionalProperties: false
      required: ["items"]
      properties:
        items:
          type: "array"
          items:
            type: "object"
            additionalProperties: false
            required: ["id", "revision", "snapshot", "reason", "publishedById", "createdAt"]
            properties:
              id:
                type: "string"
              revision:
                type: "integer"
                minimum: 1
                maximum: 2147483647
              snapshot: {}
              reason:
                type: "string"
              publishedById:
                type: "string"
              createdAt:
                type: "string"
                format: "date-time"
          maxItems: 100
...